remove all PageContext outside of Page
