+++ pages/notifications/index.tsx
@@ -6,7 +6,6 @@ import FilterButton from '../../components/filterButton';
 import NotificationList from '../../components/notification/notificationList';
 import Page from '../../components/page';
 import { AppContext } from '../../contexts/appContext';
-import { PageContext } from '../../contexts/pageContext';
 import { enrichNotifications } from '../../helpers/enrich';
 import usePush from '../../hooks/usePush';
 import dbConnect from '../../lib/dbConnect';
@@ -79,7 +78,6 @@ export default function Notifications({ notifications, searchQuery, totalRows }:
   const firstLoad = useRef(true);
   const [data, setData] = useState<Notification[]>(notifications);
   const [loading, setLoading] = useState(false);
-  const { mutateUser } = useContext(PageContext);
   const [page, setPage] = useState(searchQuery.page || 1);
   const router = useRouter();
   const routerPush = usePush();
@@ -136,7 +134,7 @@ export default function Notifications({ notifications, searchQuery, totalRows }:
         <div className='pl-3'>
           <FilterButton selected={showFilter === 'unread'} value='unread' first last onClick={onUnreadFilterButtonClick} element={<span className='text-sm'>Unread</span>} />
         </div>
-        <NotificationList mutateNotifications={mutateUser} notifications={data} setNotifications={setData} />
+        <NotificationList notifications={data} setNotifications={setData} />
         {totalRows > notificationsPerPage &&
           <div className='flex justify-center flex-row'>
             {page > 1 && (
